+++ CHANGES.rst
@@ -84,6 +84,8 @@ Imviz
 - Pressing Shift+b now blinks backwards. Right-clicking on the image while Blink tool
   is active on the toolbar also blinks backwards. [#1558]
 
+- New Simple Image Rotation plugin to rotate the axes of images. [#1551]
+
 Mosviz
 ^^^^^^